* Added new MetadataUtils * Added unit test for MetadataUtils
